before-h1

+++ src/app/component/pages/profile-page/profile-page/profile-page.component.ts
@@ -5,6 +5,7 @@ import { CountryByCallingPhone } from 'src/app/models/country-by-calling-phone';
 import { Observable, Subscription } from 'rxjs';
 import { GetHttpService } from 'src/app/service/get-http/get-http.service';
 import { graphqlService } from 'src/app/service/graphql/graphql.service';
+import { element } from 'protractor';
 
 @Component({
   selector: 'app-profile-page',
@@ -20,11 +21,14 @@ export class ProfilePageComponent implements OnInit {
   city = new FormControl()
   address = new FormControl()
   postcode = new FormControl()
-
+  validNumber: any
+  language: any [] = [
+    "ID", "EN", "CN"
+  ]
+  index = 0
   countryCode : string = ""
   isLogin : boolean = false;
   countryCode$: Observable<CountryByCallingPhone[]>;
-
   user: User = JSON.parse(localStorage.getItem('currentUser'))
 
   profiles : any = [
@@ -40,7 +44,10 @@ export class ProfilePageComponent implements OnInit {
     {"logo": "K", "desc": "Keluar"},
   ]
 
+  languageSelected : string
+
   constructor(private getHttpService : GetHttpService, private graphql: graphqlService) {
+    this.languageSelected = this.user[0].language
   }
 
   ngOnInit() {
@@ -67,21 +74,53 @@ export class ProfilePageComponent implements OnInit {
 
   updateProfile(){
     var user : User = new User()
-    this.graphql.updateUser(this.user[0].id, this.firstName.value, this.lastName.value, this.phoneNumber.value,this.email.value, this.city.value, this.address.value, this.postcode.value).subscribe(query => {
+    console.log(this.phoneNumber.value)
+    this.getHttpService.getValidNumber("http://apilayer.net/api/validate?access_key=456fb23c31b6966dfda380f8f928e5fd&number="+this.phoneNumber.value+"&country_code=&format=1").map((data) => JSON.stringify(data)).subscribe((data) => {
+      this.validNumber = JSON.parse(data).valid
+      return this.validNumber
     })
-    localStorage.clear()
-    user.lastName = this.lastName.value
-    user.email = this.email.value
-    user.phoneNumber = this.phoneNumber.value
-    user.cityName = this.city.value
-    user.address = this.address.value
-    user.postCode = this.postcode.value
-    localStorage.setItem('currentUser', JSON.stringify(user))
-  }
 
-  logout(){
+    if(!this.checkLength(this.firstName.value, 5, 10))alert("First Name must be 5 - 10 Chara")
+    if(!this.checkLength(this.lastName.value, 5, 10))alert("Last Name must be 5 - 10 Chara")
+    if(!this.validNumber)alert("Phone Format Wrong")
+    if(this.emailValidation(this.email.value))alert("Email Format Wrong")
+
+    if(this.emailValidation(this.email.value) || !this.validNumber || !this.checkLength(this.firstName.value, 5, 10) || !this.checkLength(this.lastName.value, 5, 10)) alert("Registration Failed")
+    else{
+        this.graphql.updateUser(this.user[0].id, this.firstName.value, this.lastName.value, this.phoneNumber.value,this.email.value, this.city.value, this.address.value, this.postcode.value, this.languageSelected).subscribe(async query => {
+          await
+          alert("Update Success")
+          localStorage.clear()
+          user.lastName = this.lastName.value
+          user.email = this.email.value
+          user.phoneNumber = this.phoneNumber.value
+          user.cityName = this.city.value
+          user.address = this.address.value
+          user.postCode = this.postcode.value
+          user.language = this.languageSelected
+          localStorage.setItem('currentUser', JSON.stringify(user))
+        })
+      }
+    }
+
+    logout(){
     localStorage.clear()
     window.location.reload()
   }
 
+
+  checkLength(string, min, max){
+    if(string.length <= max && string.length >= min)
+      return true
+    return false
+
+  }
+
+  emailValidation(email){
+    if(email[email.indexOf('@') - 1] ||email[email.indexOf('@') + 1])
+      return false
+    else if(!email.endsWith(".com"))
+      return false
+    return true
+  }
 }
